bug fixed- dealing with multiple ways to fix the bug of dealing with upper cases and lower cases
in this assignment ( the first one) we have found the way to deal with the bug of the caps and lowercases letters. we have found that in order that the programm will work as it should I had to write the three types of writings that I`ve wanted to check. Lowercases, capital letters and Upper cases. and then I`ve written the commands of IF and ELSE about those three conditions and finally we got the output of true\false according to the way we have written it